Split large 'Util' class into smaller, more specific classes to improve readability and maintenance
The 'Util' class was too large and was doing too many things, making it difficult to maintain and understand. Therefore, I decided to split it up into smaller classes each handling their respective utilities. Now, we have 'UiUtils', 'DialogUtils', 'ImageUtils', 'CultureUtils', and changes in 'FileUtils', each handling UI, Dialog, Image, Culture and File utilities respectively. This enhances readability of the code and makes it easier for future changes or bug fixes.
