+++ FoliCon/Modules/IGDBClass.cs
@@ -108,7 +108,7 @@ public class IgdbClass
         var localPosterPath = fullFolderPath + @"\" + folderName + ".png";
         var year = result.FirstReleaseDate != null ? result.FirstReleaseDate.Value.Year.ToString(CultureInfo.InvariantCulture) : "";
         var posterUrl = ImageHelper.GetImageUrl(result.Cover.Value.ImageId, ImageSize.HD720);
-        Util.AddToPickedListDataTable(_listDataTable, localPosterPath, result.Name, rating, fullFolderPath, folderName,
+        FileUtils.AddToPickedListDataTable(_listDataTable, localPosterPath, result.Name, rating, fullFolderPath, folderName,
             year);
         if (result.Id != null) FileUtils.SaveMediaInfo((int)result.Id, "Game", fullFolderPath);
         var tempImage = new ImageToDownload
